requirements.txt
Django
django-heroku
gunicorn
requests
behave-django